chore(seo): extract PageMeta, hoist into App, drop live price from title
Move the inline PageMeta out of components/Layout/Page.tsx into its own
file, then mount it once at the top of the Router in App.tsx so every
route gets meta regardless of which Page wrapper it uses (the old
behaviour coupled meta to views/Page.tsx, which has no Helmet, so
/swap /add /remove /find /liquidity all silently lost meta).

Also delete Home/index.tsx's duplicate <Helmet> block — it bypassed
getCustomMeta and re-injected the live $PLANT price into <title>,
producing unstable SERP snippets like 'Swap | PlantSwap - $0.123'.
Home now uses the hoisted PageMeta with the price-free title.
